% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{get_stations}
\alias{get_stations}
\title{Get bike-sharing stations}
\usage{
get_stations(network_id)
}
\arguments{
\item{network_id}{The ID of the CityBikes network.}
}
\value{
A data frame containing station names, locations,
  update times, available bikes, and empty slots.
}
\description{
Gets information about the stations in a CityBikes network,
including their location and current bike availability.
}
